chore : calender 함수 이름 변경 및 콘솔로그 제거
